Separate pl visualizers into separate files
