+++ main/gui/source/init/tasks.cpp
@@ -276,7 +276,7 @@ namespace hex::init {
         };
 
         // Glyph range for blender icons
-        static ImWchar blenderIconsRange[] = {
+        static std::array<ImWchar, 3> blenderIconsRange = {
                 ICON_MIN_BI, ICON_MAX_BI, 0
         };
 
@@ -310,22 +310,12 @@ namespace hex::init {
         // Merge all fonts into one big font atlas
         cfg.MergeMode = true;
 
-        // Add font awesome and codicons icons to font atlas
-        cfg.FontBuilderFlags |= ImGuiFreeTypeBuilderFlags_Monochrome;
-        cfg.FontBuilderFlags |= ImGuiFreeTypeBuilderFlags_MonoHinting;
-        cfg.GlyphOffset = ImVec2(0, -2_scaled);
-        fonts->AddFontFromMemoryCompressedTTF(font_awesome_compressed_data, font_awesome_compressed_size, 0, &cfg, fontAwesomeRange);
-
-        cfg.GlyphOffset = ImVec2(0, 0);
-        //cfg.SizePixels = fontSize;
-        //fonts->AddFontFromFileTTF(wolv::util::toUTF8String(fontFile).c_str(), 0, &cfg, ranges2.Data);
-        fonts->AddFontFromMemoryCompressedTTF(codicons_compressed_data, codicons_compressed_size, 0, &cfg, codiconsRange);
-
-        cfg.GlyphOffset = ImVec2(0, -3_scaled);
-        fonts->AddFontFromMemoryTTF((void *) blendericons_data, blendericons_size, 0, &cfg, blenderIconsRange);
-        //fonts->AddFontFromMemoryCompressedTTF(blender_icons_compressed_data, blender_icons_compressed_size, 0, &cfg, blenderIconsRange);
-        cfg.FontBuilderFlags &= ~ImGuiFreeTypeBuilderFlags_Monochrome;
-        cfg.FontBuilderFlags &= ~ImGuiFreeTypeBuilderFlags_MonoHinting;
+        cfg.GlyphOffset = ImVec2(1_scaled, 1_scaled);
+        fonts->AddFontFromMemoryCompressedTTF(font_awesome_compressed_data, font_awesome_compressed_size, 0, &cfg, fontAwesomeRange.data());
+        cfg.GlyphOffset = ImVec2(0, 3_scaled);
+        fonts->AddFontFromMemoryCompressedTTF(codicons_compressed_data, codicons_compressed_size, 0, &cfg, codiconsRange.data());
+        cfg.GlyphOffset = ImVec2(1_scaled, 0_scaled);
+        fonts->AddFontFromMemoryTTF((void*)blendericons_data, blendericons_size, 0, &cfg, blenderIconsRange.data());
 
         cfg.GlyphOffset = ImVec2(0, 0);
         // Add unifont if unicode support is enabled
